feat: extract components to a file
